boss2移植　基本機能のみ
